D2/Ex2/requirement.txt
google-cloud-bigquery
pyarrow
pandas
google-cloud-bigquery-storage